[Patch v2] Add onError callback for stdin 'error' and 'end'
Previously prompt would only listen for 'keypress' events, which meant
that if the input ended (either by reaching the end of an input file or
pipe, or the user sending EOF via ^D at the terminal) prompt would never
return.  This commit adds an onError callback which is called if stdin
emits 'error' or 'end' events which signal the end of input.

Update code, docs, and tests to handle this:
- Update password and multi to accept an onError callback and pass it to
  prompt.
- Add it to the name example so that users know how to use it.
- Add a test to check that the example behaves as expected.

Changes in v2:
- Add onError callback for end/error instead of returning a null value.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>